# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "econiq domain API",
    "description": "Evidence-backed, Process-first investment research.\n\nThe chain every response can be walked back along:\n\n    Documents \u2192 Claims \u2192 Events \u2192 Processes \u2192 States\n    \u2192 Bottlenecks \u2192 Capabilities \u2192 Assets\n\nReads accept `as_of` to reconstruct the graph as it was at a past instant.\nScores are returned per family with their inputs attached \u2014 Thesis quality,\nAsset quality and Trade quality are never combined.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/processes": {
      "get": {
        "tags": [
          "processes"
        ],
        "summary": "List Processes",
        "operationId": "list_processes_api_processes_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "requires_review",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requires Review"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then.",
              "title": "As Of"
            },
            "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessSummaryOut"
                  },
                  "title": "Response List Processes Api Processes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/processes/{process_id}": {
      "get": {
        "tags": [
          "processes"
        ],
        "summary": "Get Process",
        "operationId": "get_process_api_processes__process_id__get",
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Process Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then.",
              "title": "As Of"
            },
            "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/processes/{process_id}/states": {
      "get": {
        "tags": [
          "processes"
        ],
        "summary": "State History",
        "description": "Append-only State history \u2014 the sequence *is* the record (ui_concept \u00a732).",
        "operationId": "state_history_api_processes__process_id__states_get",
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Process Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then.",
              "title": "As Of"
            },
            "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessStateOut"
                  },
                  "title": "Response State History Api Processes  Process Id  States Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/processes/{process_id}/timeline": {
      "get": {
        "tags": [
          "processes"
        ],
        "summary": "Timeline",
        "description": "State changes, belief changes, evidence and critiques on one axis (#20).\n\nThree separate lists would leave the reader joining them by eye, and the\njoin is the point: a belief change is only defensible next to the evidence\nthat arrived just before it. The entries carry both dates \u2014 when the thing\nhappened and when the system learned it \u2014 because a document published in\nJuly and ingested in August belongs in two different places depending on\nwhich question is being asked.",
        "operationId": "timeline_api_processes__process_id__timeline_get",
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Process Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then.",
              "title": "As Of"
            },
            "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessTimelineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/processes/{process_id}/journal": {
      "get": {
        "tags": [
          "processes"
        ],
        "summary": "Journal",
        "description": "Why the system changed its mind, chronologically (PRD \u00a721).",
        "operationId": "journal_api_processes__process_id__journal_get",
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Process Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then.",
              "title": "As Of"
            },
            "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalEntryOut"
                  },
                  "title": "Response Journal Api Processes  Process Id  Journal Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/processes/{process_id}/critiques": {
      "get": {
        "tags": [
          "processes"
        ],
        "summary": "Critiques",
        "description": "Adversarial findings, including superseded ones.\n\nAddressed critiques are returned rather than hidden: a thesis that survived\nfour attacks is stronger than one never attacked, and that is only visible\nif the attacks remain on the record.",
        "operationId": "critiques_api_processes__process_id__critiques_get",
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Process Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CritiqueStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then.",
              "title": "As Of"
            },
            "description": "Reconstruct the graph as it was at this instant. Omit for the current state. Revisions and observations recorded later are excluded, so a replay describes what was believed then."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CritiqueOut"
                  },
                  "title": "Response Critiques Api Processes  Process Id  Critiques Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessStatus": {
        "type": "string",
        "enum": [
          "candidate",
          "active",
          "dormant",
          "invalidated",
          "merged",
          "concluded"
        ],
        "title": "ProcessStatus",
        "description": "Lifecycle of the Process object itself, distinct from its economic\nState (ontology \u00a77 vs \u00a710)."
      },
      "ProcessSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "archetype": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessArchetype"
              },
              {
                "type": "null"
              }
            ]
          },
          "archetype_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archetype Confidence"
          },
          "status": {
            "$ref": "#/components/schemas/ProcessStatus"
          },
          "requires_review": {
            "type": "boolean",
            "title": "Requires Review"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "current_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessStateLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "state_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Confidence"
          },
          "state_observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Observed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "archetype",
          "archetype_confidence",
          "status",
          "requires_review",
          "revision"
        ],
        "title": "ProcessSummaryOut"
      },
...
    }
  }
}